# Speed up SkyCoord initialization by a factor of 2-3