--- a/astropy/coordinates/baseframe.py
+++ b/astropy/coordinates/baseframe.py
@@ -766,9 +766,13 @@ class BaseCoordinateFrame(ShapedLikeNDArray, metaclass=FrameMeta):
 
     @classmethod
     def _get_representation_info(cls):
+
         # This exists as a class method only to support handling frame inputs
         # without units, which are deprecated and will be removed.  This can be
         # moved into the representation_info property at that time.
+
+        if hasattr(cls, '_representation_info_cache'):
+            return cls._representation_info_cache
 
         repr_attrs = {}
         for repr_diff_cls in (list(r.REPRESENTATION_CLASSES.values()) +
@@ -805,6 +809,8 @@ class BaseCoordinateFrame(ShapedLikeNDArray, metaclass=FrameMeta):
             # Convert to tuples so that this can't mess with frame internals
             repr_attrs[repr_diff_cls]['names'] = tuple(nms)
             repr_attrs[repr_diff_cls]['units'] = tuple(uns)
+
+        cls._representation_info_cache = repr_attrs
 
         return repr_attrs