This implements the caching of ``representation_info`` on frames on a class level to speed up initialization of SkyCoords (as described in https://github.com/astropy/astropy/issues/2989)

### Before

```python
In [2]: %timeit SkyCoord(1, 2, unit='deg', frame='icrs')
1.15 ms ± 54.3 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [5]: ra = np.random.random(1000) * u.deg

In [6]: dec = np.random.random(1000) * u.deg

In [7]: %timeit SkyCoord(ra, dec)
1.17 ms ± 51.5 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

### After

```python
In [2]: %timeit SkyCoord(1, 2, unit='deg', frame='icrs')
494 µs ± 63.7 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [6]: ra = np.random.random(1000) * u.deg

In [7]: dec = np.random.random(1000) * u.deg

In [8]: %timeit SkyCoord(ra, dec)
456 µs ± 24.6 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

There is a speedup even when initializing a single ``SkyCoord`` because ``_get_representation_info`` gets called 10 times for each ``SkyCoord`` initialization, so the caching helps even for a single initialization.

I think in general there are other things in frames that could be classed at a class level, for example ``representation_component_names`` and ``representation_component_units``. I don't have time to look into other optimizations, but this one was bugging me since it's such a simple fix.