@astrofrog 2018-07-01
Ok so there is way too much class magic going on in astropy.coordinates, and this doesn't work. See https://github.com/astropy/astropy/pull/7615 for caching on an instance level which still provides a factor of ~2 improvement.

I really think it would be good to tidy up the frame code so that things can be computed once and for all for each frame, and I think there is a fair bit of opportunity there for optimization, but I don't have the time to do this.

I'll leave this open for @eteq and @adrn to see that it doesn't work, but feel free to close.